Move path finding/seletion to its own package
Add paths_gen package and put all path generation related stuff there.

Commit: Ref
