+++ src/paths_gen/PathFinder.java
@@ -1,4 +1,4 @@
-package rbr;
+package paths_gen;
 
 import util.*;
 import java.util.*;
